feat: adjust & remove fields
There is no way to change the user's `star_friend` and `remark_name`, which are all fixed to empty value, so I only reserve them for the contact class.
All contacts' `app_account_flag` are always 0, but I'm not sure the user's is also the same. Instead of removing it, I reserve this field for the user class.
Since the `owner_uin`, `uni_friend` and `chat_room_id` of all contacts are just empty, I remove them with faith.